Fix timeout and backward compatibility issues #52 (#1)
* Fix issue with large usage of time to find and select an option [Fixes #1]

Before this commit #select2 checked if there is single select box on the
page using has_selector?('.select2-choise'). It caused capybara to wait
default wait time in case when we test multiple select box. This often
leads scenarios to fail with timeout error and slow tests

This commit tries to avoid using has_selector?

* Click on the search field to open multiple select box

Because sometimes it fails to open when clicking on choises if there are many

* Fix backward compatibility with select2 <v4.0 when search is specified
[Fix #51]